requirements.txt
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cycler==0.11.0
DateTime==4.3
et-xmlfile==1.1.0
Flask==2.0.2
fonttools==4.29.1
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pdfrw==0.4
Pillow==9.0.1
pyparsing==3.0.7
PyPDF2==1.26.0
pyshp==2.2.0
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.7
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
typing_extensions==4.0.1
waitress==2.1.2
Werkzeug==2.0.2
xlwt==1.3.0
zipp==3.7.0
zope.interface==5.4.0
